+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
-* `edgeToEdge {}` can be applied multiple times.
+* New `Fragment.fitEdgeToEdge()` function forcing the view previously configured in `edgeToEdge {}` block to fit the edges. 
 * New `ConstraintLayout + Transition` sample screen.
+* `edgeToEdge {}` can be applied multiple times.
 
 # 0.4 (29.12.2019)
 * New public `Window.setEdgeToEdgeFlags()`.